some more adjustments for large mesh

+++ inc/defines.hpp
@@ -80,8 +80,8 @@ typedef fpml::fixed_point<int32_t,31-FXP_FRACPART,FXP_FRACPART> fixp32;
 #endif
 typedef fpml::fixed_point<int64_t,63-FXP_FRACPART,FXP_FRACPART> fixp64;
 
-// has to be uint32_t for OpenCL
-typedef uint32_t meshindex_t;
+// check OpenCL code when changing
+typedef uint64_t meshindex_t;
 
 typedef float meshaxis_t;
 typedef float meshdata_t;
